Create DOM Class
DOM Class is a mixture of prior functions that
were implemented on index.js and DOM.js.
Seperate audits and eliminate much of the tight
coupling from functions.
Completely Refactor index.js

The file was becoming a nightmare to debug, and
I was having trouble storing an updated version of
validArray until implementing a class. Index.js is
composed almost entirely of event listeners.